Don't constantly reload pages we don't need to
This improves the performance of pages such as bones which
loads an image from the web somewhere, as it doesn't need to
be loaded every min.
This will also decrease the # of times that the lcpt api will be called
to fetch events.

The page as a whole will still be reloaded every 15 mins in order
to make sure that the whole thing is pretty fresh.

If we do ever add screens such as a twitter feed or latest tweet
that need to be updated more regularly then the list of URLs could be
converted to objects containing more info such as url, shouldReload etc.

Another small improvement could be made here making sure that the iframe
is actually fully loaded before switching to do (avoiding any blank time
on the screens at all)